+++ bugsnag-plugin-android-ndk/src/main/jni/handlers/cpp_handler.cpp
@@ -110,6 +110,7 @@ void bsg_handle_cpp_terminate() {
   if (bsg_run_on_error()) {
     bsg_increment_unhandled_count(&bsg_global_env->next_event);
     bsg_serialize_event_to_file(bsg_global_env);
+    bsg_serialize_last_run_info_to_file(bsg_global_env);
   }
   bsg_global_env->crash_handled = true;
   bsg_handler_uninstall_cpp();